Fix word splitting issues
This changes eliminates several issues with paths containing spaces.
First of all variables are now properly quoted. Then iteration over find
output takes an intermediate step with a temporary file to avoid word
splitting in a path.

The call to head -c 1 is replaced by printf, which is probably a shell
built-in and thus might save an external process call.

The outer iteration over the result of extract has moved from the
create_dir function to a main function.